+++ app/src/main/java/org/scottishtecharmy/soundscape/screens/markers_routes/screens/MarkersAndRoutesScreen.kt
@@ -15,7 +15,7 @@ import org.scottishtecharmy.soundscape.screens.home.HomeRoutes
 import org.scottishtecharmy.soundscape.screens.markers_routes.components.MarkersAndRoutesAppBar
 import org.scottishtecharmy.soundscape.screens.markers_routes.components.MarkersAndRoutesTabs
 import org.scottishtecharmy.soundscape.screens.markers_routes.marker_route_screens.RoutesScreen
-import org.scottishtecharmy.soundscape.screens.markers_routes.navigation.ScreensForMarkersAndRoutes
+import org.scottishtecharmy.soundscape.screens.markers_routes.screens.markersscreen.MarkersScreen
 import org.scottishtecharmy.soundscape.ui.theme.SoundscapeTheme
 
 @Composable
