+++ app/src/main/java/org/scottishtecharmy/soundscape/screens/markers_routes/screens/SharedPreferences.kt
@@ -1,4 +1,4 @@
-package org.scottishtecharmy.soundscape.screens.markers_routes
+package org.scottishtecharmy.soundscape.screens.markers_routes.screens
 
 import android.content.Context
